[code & style]: update format for .2f & reorganize the file organization
